[BUGFIX] Extract __version__ to safe import module
The `trakt/__init__.py` references many modules including 3rd party
modules which might not be installed already. The import within the
setup.py results in an ImportError.

fixes #40
