+++ trakt/__init__.py
@@ -1,7 +1,8 @@
 from trakt.core.errors import ERRORS
 from trakt.core.exceptions import RequestError, ClientError, ServerError
-from trakt.client import TraktClient, __version__
+from trakt.client import TraktClient
 from trakt.helpers import has_attribute
+from trakt.version import __version__
 
 from six import add_metaclass
